# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing community waste collection — households, pickups, payments, reports.",
        "title": "Community Waste Collection API",
        "contact": {
            "name": "Muhammad Faisal Affan"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/pickups": {
            "get": {
                "description": "List pickups with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pickups"
                ],
                "summary": "List pickups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by household",
                        "name": "household_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a waste pickup request (BR-01: blocks if pending payment exists)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pickups"
                ],
                "summary": "Create pickup request",
                "parameters": [
                    {
                        "description": "Pickup data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.CreatePickupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.WastePickup"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Pending payment exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/pickups/{id}/cancel": {
            "put": {
                "description": "Cancel a pending pickup",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pickups"
                ],
                "summary": "Cancel pickup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickup UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.WastePickup"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/pickups/{id}/complete": {
            "put": {
                "description": "Mark pickup as completed (BR-05: auto-generates payment)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pickups"
                ],
                "summary": "Complete pickup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickup UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/pickups/{id}/schedule": {
            "put": {
                "description": "Schedule a pending pickup (BR-02, BR-03: safety check for electronic)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pickups"
                ],
                "summary": "Schedule pickup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickup UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.SchedulePickupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_internal_domain.WastePickup"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
        "/reports/households/{id}/history": {
            "get": {
                "description": "Full pickup and payment history for a household",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Household history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Household UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.CreatePickupRequest": {
            "type": "object",
            "required": [
                "household_id",
                "type"
            ],
            "properties": {
                "household_id": {
                    "type": "string"
                },
                "safety_check": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "organic",
                        "plastic",
                        "paper",
                        "electronic"
                    ]
                }
            }
        },
...
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.SchedulePickupRequest": {
            "type": "object",
            "required": [
                "pickup_date"
            ],
            "properties": {
                "pickup_date": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_internal_domain.WastePickup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "household_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pickup_date": {
                    "type": "string"
                },
                "safety_check": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.apiError"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.pagination"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.ValidationDetail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.apiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_faisalaffan_community-waste-collection-api_pkg_response.ValidationDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_faisalaffan_community-waste-collection-api_pkg_response.pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        }
    }
}